+++ src/Tests.cpp
@@ -6140,6 +6140,153 @@ static void TestMemoryUsage()
     }
 }
 
+static void TestAllocationWithAlignment()
+{
+    wprintf(L"Test allocation with alignment\n");
+
+    static const VkDeviceSize BUFFER_SIZE = 4 * KILOBYTE;
+    static const VkDeviceSize MIN_ALIGNMENT = 64 * KILOBYTE;
+
+    VkBufferCreateInfo bufCreateInfo = { VK_STRUCTURE_TYPE_BUFFER_CREATE_INFO };
+    bufCreateInfo.size = BUFFER_SIZE;
+    bufCreateInfo.usage = VK_BUFFER_USAGE_TRANSFER_DST_BIT | VK_BUFFER_USAGE_TRANSFER_SRC_BIT;
+    
+    VkResult res;
+
+    // 1. Using vmaAllocateMemory with VmaAllocationCreateInfo::minAlignment
+    {
+        VkBuffer buffers[2] = {};
+        VkMemoryRequirements memReq[2] = {};
+        VmaAllocation allocations[2] = {};
+        VmaAllocationInfo allocInfo[2] = {};
+
+        for(uint32_t i = 0; i < 2; ++i)
+        {
+            res = vkCreateBuffer(g_hDevice, &bufCreateInfo, g_Allocs, &buffers[i]);
+            TEST(res == VK_SUCCESS && buffers[i] != VK_NULL_HANDLE);
+            vkGetBufferMemoryRequirements(g_hDevice, buffers[i], &memReq[i]);
+        }
+
+        VmaAllocationCreateInfo allocCreateInfo = {};
+        allocCreateInfo.minAlignment = MIN_ALIGNMENT;
+
+        for(uint32_t i = 0; i < 2; ++i)
+        {
+            res = vmaAllocateMemory(g_hAllocator, &memReq[i], &allocCreateInfo, &allocations[i], &allocInfo[i]);
+            TEST(res == VK_SUCCESS && allocations[i] != VK_NULL_HANDLE);
+            TEST(allocInfo[i].offset % MIN_ALIGNMENT == 0); // !!!
+        }
+
+        for(uint32_t i = 0; i < 2; ++i)
+        {
+            res = vmaBindBufferMemory(g_hAllocator, allocations[i], buffers[i]);
+            TEST(res == VK_SUCCESS);
+        }
+
+        for(uint32_t i = 0; i < 2; ++i)
+        {
+            vkDestroyBuffer(g_hDevice, buffers[i], g_Allocs);
+            vmaFreeMemory(g_hAllocator, allocations[i]);
+        }
+    }
+
+    // 2. Using vmaCreateBuffer with VmaAllocationCreateInfo::minAlignment
+    {
+        VkBuffer buffers[2] = {};
+        VmaAllocation allocations[2] = {};
+        VmaAllocationInfo allocInfo[2] = {};
+
+        VmaAllocationCreateInfo allocCreateInfo = {};
+        allocCreateInfo.minAlignment = MIN_ALIGNMENT;
+
+        for(uint32_t i = 0; i < 2; ++i)
+        {
+            res = vmaCreateBuffer(g_hAllocator, &bufCreateInfo, &allocCreateInfo, &buffers[i], &allocations[i], &allocInfo[i]);
+            TEST(res == VK_SUCCESS && buffers[i] != VK_NULL_HANDLE && allocations[i] != VK_NULL_HANDLE);
+            TEST(allocInfo[i].offset % MIN_ALIGNMENT == 0); // !!!
+        }
+
+        for(uint32_t i = 0; i < 2; ++i)
+        {
+            vmaDestroyBuffer(g_hAllocator, buffers[i], allocations[i]);
+        }
+    }
+
+    // 3. Using vmaCreateBuffer in a custom pool with VmaPoolCreateInfo::minAllocationAlignment specified
+    {
+        VmaAllocationCreateInfo sampleAllocCreateInfo = {};
+        sampleAllocCreateInfo.usage = VMA_MEMORY_USAGE_AUTO_PREFER_HOST;
+
+        VmaPoolCreateInfo poolCreateInfo = {};
+        res = vmaFindMemoryTypeIndexForBufferInfo(g_hAllocator, &bufCreateInfo, &sampleAllocCreateInfo, &poolCreateInfo.memoryTypeIndex);
+        TEST(res == VK_SUCCESS);
+
+        poolCreateInfo.blockSize = 4 * MIN_ALIGNMENT;
+        poolCreateInfo.minBlockCount = 1;
+        poolCreateInfo.maxBlockCount = 1;
+        poolCreateInfo.minAllocationAlignment = MIN_ALIGNMENT;
+
+        VmaPool pool = VK_NULL_HANDLE;
+        if(res == VK_SUCCESS)
+        {
+            res = vmaCreatePool(g_hAllocator, &poolCreateInfo, &pool);
+            TEST(res == VK_SUCCESS && pool != VK_NULL_HANDLE);
+        }
+
+        VkBuffer buffers[2] = {};
+        VmaAllocation allocations[2] = {};
+        VmaAllocationInfo allocInfo[2] = {};
+
+        VmaAllocationCreateInfo allocCreateInfo = {};
+        allocCreateInfo.pool = pool;
+
+        if(pool != VK_NULL_HANDLE)
+        {
+            for(uint32_t i = 0; i < 2; ++i)
+            {
+                res = vmaCreateBuffer(g_hAllocator, &bufCreateInfo, &allocCreateInfo, &buffers[i], &allocations[i], &allocInfo[i]);
+                TEST(res == VK_SUCCESS && buffers[i] != VK_NULL_HANDLE && allocations[i] != VK_NULL_HANDLE);
+                TEST(allocInfo[i].offset % MIN_ALIGNMENT == 0); // !!!
+            }
+        }
+
+        for(uint32_t i = 0; i < 2; ++i)
+        {
+            vmaDestroyBuffer(g_hAllocator, buffers[i], allocations[i]);
+        }
+
+        vmaDestroyPool(g_hAllocator, pool);
+    }
+
+    // 4. Using vmaCreateBufferWithAlignment
+    {
+        VkBuffer buffers[2] = {};
+        VmaAllocation allocations[2] = {};
+        VmaAllocationInfo allocInfo[2] = {};
+
+        VmaAllocationCreateInfo allocCreateInfo = {};
+
+        for(uint32_t i = 0; i < 2; ++i)
+        {
+            res = vmaCreateBufferWithAlignment(
+                g_hAllocator,
+                &bufCreateInfo,
+                &allocCreateInfo,
+                MIN_ALIGNMENT,
+                &buffers[i],
+                &allocations[i],
+                &allocInfo[i]);
+            TEST(res == VK_SUCCESS && buffers[i] != VK_NULL_HANDLE && allocations[i] != VK_NULL_HANDLE);
+            TEST(allocInfo[i].offset % MIN_ALIGNMENT == 0); // !!!
+        }
+
+        for(uint32_t i = 0; i < 2; ++i)
+        {
+            vmaDestroyBuffer(g_hAllocator, buffers[i], allocations[i]);
+        }
+    }
+}
+
 static void TestDataUploadingWithStagingBuffer()
 {
     wprintf(L"Testing data uploading with staging buffer...\n");
@@ -8727,6 +8874,7 @@ void Test()
     TestAllocationsInitialization();
 #endif
     TestMemoryUsage();
+    TestAllocationWithAlignment();
     TestDataUploadingWithStagingBuffer();
     TestDataUploadingWithMappedMemory();
     TestAdvancedDataUploading();